correct setLockTargetState in readme
The biggest problem is that API item 2 had the wrong API call in the documentation. But more generally, the documentation is clearer using the words lock/unlock instead of open/close, since a lock really does not open or close anything. Some locks do have a sensor to know when a door is actually open or closed, and the auto lock on these devices may be related to that open/close state. But this plugin is not really aware of the open/close state of the locked item, so it is best to use locked/unlocked consistently.